Resolve clippy warnings and use correct types in bindings.
bindings were using i8 instead of c_char for C string pointers, but
the types are defined as c_char in libkrun.

There were a lot of unecessary casts from i8 to c_char according to clippy
because on x86_64 they are the same type.
(The casts were necessary on ARM, because c_char is unsigned on ARM).

Signed-off-by: Matej Hrica <mhrica@redhat.com>