Add support for starting VMs with passt network
This required adding another field to the config. This is done by
migrating the old config to a newer version.
We are backwards compatible on configuration but not forwards compatible
(older versions of krunvm will not be able to use the config from
this version)
If we want forward compatibility, I feel like we need to ditch the confy
crate.

Signed-off-by: Matej Hrica <mhrica@redhat.com>