start: Only map volumes if required on macOS
In issue #67, it was reported that VM creation on macOS always fails
with the message: "Error setting VM mapped volumes". [^1] It appears
krun always calls `krun_set_mapped_volumes` even if the user has not
configured any volumes. That krun library function was disabled in
Feb 2024 by commit 186083df9. [^2] Based on the commit message, it
sounds like virtio-fs devices will be the preferred way to implement
mapped volumes going forward.

For now, we can at least avoid calling the deprecated function if
the user has not configured any mapped volumes. That will obviously
not fix the krun feature, but it makes krunvm usable on macOS.

References

1. https://github.com/containers/krunvm/issues/67
2. libkrun commit 186083df9faf704e2b431d90b1ed147363aca004

Signed-off-by: Stephen Fox <stephen.j.fox.jr@gmail.com>