Re-execute itself with DYLD_LIBRARY_PATH
After libkrun stopped linking directly against libkrunfw, we depend on
dlopen() and the system's linker to do a sane job at looking for the
library. This poses a problem on macOS, since the library path from
the brew prefix is not used for looking up libraries by default.

So work around this limitation, look if the DYLD_LIBRARY_PATH
environment variable is present. If it isn't, obtain the brew prefix and
re-execute krunvm with pushing that variable with brew's library path
into the list of environment variables.

Fixes: #70
Signed-off-by: Sergio Lopez <slp@redhat.com>